+++ saleor/app/tests/test_installation_utils.py
@@ -4,7 +4,11 @@ from unittest.mock import ANY, Mock, patch
 import graphene
 import pytest
 import requests
+from celery.exceptions import Retry
 from django.core.exceptions import ValidationError
+from django.core.files import File
+from django.core.files.base import ContentFile
+from django.db import DatabaseError
 from freezegun import freeze_time
 
 from ... import __version__
@@ -13,7 +17,10 @@ from ...webhook.event_types import WebhookEventAsyncType
 from ...webhook.payloads import generate_meta, generate_requestor
 from ..error_codes import AppErrorCode
 from ..installation_utils import (
+    MAX_ICON_FILE_SIZE,
     AppInstallationError,
+    fetch_brand_data_task,
+    fetch_icon_image,
     install_app,
     validate_app_install_response,
 )
@@ -166,6 +173,29 @@ def test_install_app_with_empty_author(app_manifest, app_installation, monkeypat
     errors = validation_error.value.error_dict["author"]
     assert len(errors) == 1
     assert errors[0].code == AppErrorCode.INVALID.value
+
+
+def test_install_app_with_brand_data(app_manifest, app_installation, monkeypatch):
+    # given
+    brand_data = {"logo": {"default": "https://example.com/logo.png"}}
+    app_manifest["brand"] = brand_data
+    mocked_get_response = Mock()
+    mocked_get_response.json.return_value = app_manifest
+    monkeypatch.setattr(requests, "get", Mock(return_value=mocked_get_response))
+    monkeypatch.setattr("saleor.app.installation_utils.send_app_token", Mock())
+    mocked_fetch_brand_data_task = Mock()
+    monkeypatch.setattr(
+        "saleor.app.installation_utils.fetch_brand_data_task.delay",
+        mocked_fetch_brand_data_task,
+    )
+
+    # when
+    app, _ = install_app(app_installation, activate=True)
+
+    # then
+    mocked_fetch_brand_data_task.assert_called_once_with(
+        brand_data, app_installation_id=None, app_id=app.id
+    )
 
 
 @freeze_time("2022-05-12 12:00:00")
@@ -575,3 +605,202 @@ def test_install_app_lack_of_token_target_url_in_manifest_data(
     error_dict = excinfo.value.error_dict
     assert "tokenTargetUrl" in error_dict
     assert error_dict["tokenTargetUrl"][0].message == "Field required."
+
+
+@pytest.fixture
+def image_response_mock():
+    content_chunks = [b"fake ", b"image ", b"content"]
+    mock_response = Mock()
+    mock_response.status_code = 200
+    mock_response.content = b"".join(content_chunks)
+    mock_response.headers = {
+        "content-type": "image/png",
+        "content-length": str(len(mock_response.content)),
+    }
+    mock_response.iter_content.return_value = iter(content_chunks)
+    return mock_response
+
+
+@patch("saleor.app.installation_utils.validate_icon_image")
+@patch("saleor.app.installation_utils.requests.get")
+def test_fetch_icon_image(
+    mock_get_request, mock_validate_icon_image, image_response_mock
+):
+    # given
+    image_file_format = "png"
+    image_url = f"https://example.com/logo.{image_file_format}"
+    mock_get_request.return_value.__enter__.return_value = image_response_mock
+
+    # when
+    image_file = fetch_icon_image(image_url)
+
+    # then
+    mock_get_request.assert_called_once_with(
+        image_url, stream=True, timeout=ANY, allow_redirects=False
+    )
+    mock_validate_icon_image.assert_called_once_with(image_file, ANY)
+    assert isinstance(image_file, File)
+    assert image_file.read() == image_response_mock.content
+    assert image_file.name.endswith(image_file_format)
+
+
+@patch("saleor.app.installation_utils.validate_icon_image")
+@patch("saleor.app.installation_utils.requests.get")
+def test_fetch_icon_image_invalid_type(
+    mock_get_request, mock_validate_icon_image, image_response_mock
+):
+    mock_get_request.return_value.__enter__.return_value = image_response_mock
+    image_response_mock.headers["content-type"] = "text/html"
+
+    with pytest.raises(ValidationError) as error:
+        fetch_icon_image("https://example.com/logo.png")
+    assert error.value.code == AppErrorCode.INVALID.value
+    mock_validate_icon_image.assert_not_called()
+
+
+@patch("saleor.app.installation_utils.requests.get")
+def test_fetch_icon_image_content_length(mock_get_request, image_response_mock):
+    mock_get_request.return_value.__enter__.return_value = image_response_mock
+    image_response_mock.headers["content-length"] = MAX_ICON_FILE_SIZE + 1
+
+    with pytest.raises(ValidationError) as error:
+        fetch_icon_image("https://example.com/logo.png")
+    assert error.value.code == AppErrorCode.INVALID.value
+    assert "File too big. Maximal icon image file size is" in error.value.message
+
+
+@patch("saleor.app.installation_utils.requests.get")
+def test_fetch_icon_image_file_too_big(mock_get_request, image_response_mock):
+    def content_chunks():
+        while True:
+            yield b"0" * 1024
+
+    mock_get_request.return_value.__enter__.return_value = image_response_mock
+    image_response_mock.iter_content.return_value = content_chunks()
+
+    with pytest.raises(ValidationError) as error:
+        fetch_icon_image("https://example.com/logo.png")
+    assert error.value.code == AppErrorCode.INVALID.value
+    assert "File too big. Maximal icon image file size is" in error.value.message
+
+
+@patch("saleor.app.installation_utils.requests.get")
+def test_fetch_icon_image_network_error(mock_get_request):
+    mock_get_request.side_effect = requests.RequestException
+    with pytest.raises(ValidationError) as error:
+        fetch_icon_image("https://example.com/logo.png")
+    assert error.value.code == AppErrorCode.MANIFEST_URL_CANT_CONNECT.value
+
+
+@pytest.mark.parametrize("app_object", ["app", "app_installation"])
+@patch("saleor.app.installation_utils.fetch_icon_image")
+def test_fetch_brand_data_task(
+    mock_fetch_icon_image, app_object, app_installation, app, media_root
+):
+    # given
+    logo_url = "https://example.com/logo.png"
+    fake_img_content = b"these are bytes"
+    logo_img = ContentFile(fake_img_content, "logo.png")
+    mock_fetch_icon_image.return_value = logo_img
+
+    # when
+    fetch_brand_data_task(
+        {"logo": {"default": logo_url}},
+        app_installation_id=None if app_object == "app" else app_installation.id,
+        app_id=None if app_object != "app" else app.id,
+    )
+
+    # then
+    app.refresh_from_db()
+    app_installation.refresh_from_db()
+    mock_fetch_icon_image.assert_called_once_with(logo_url)
+    if app_object == "app":
+        assert app.brand_logo_default.read() == fake_img_content
+        assert bool(app_installation.brand_logo_default) is False
+    else:
+        assert app_installation.brand_logo_default.read() == fake_img_content
+        assert bool(app.brand_logo_default) is False
+
+
+@patch("saleor.app.installation_utils.fetch_icon_image")
+def test_fetch_brand_data_task_terminated(
+    mock_fetch_icon_image, app_installation, app, media_root
+):
+    app.delete(), app_installation.delete()
+    fetch_brand_data_task({}, app_installation_id=app_installation.id, app_id=app.id)
+    mock_fetch_icon_image.assert_not_called()
+
+
+@patch("saleor.app.installation_utils.fetch_icon_image")
+def test_fetch_brand_data_task_terminated_when_brand_data_fetched(
+    mock_fetch_icon_image, app_installation, app, media_root
+):
+    app_installation.delete()
+    app.brand_logo_default.save("logo.png", ContentFile(b"bytes"))
+    fetch_brand_data_task({}, app_installation_id=app_installation.id, app_id=app.id)
+    mock_fetch_icon_image.assert_not_called()
+
+
+@patch("saleor.app.installation_utils.fetch_icon_image")
+def test_fetch_brand_data_task_retry(
+    mock_fetch_icon_image, app_installation, app, media_root
+):
+    # given
+    brand_data = {"logo": {"default": "https://example.com/logo.png"}}
+    mock_fetch_icon_image.side_effect = ValidationError("Fetch image error")
+
+    # when
+    with pytest.raises(Retry):
+        fetch_brand_data_task(
+            brand_data, app_installation_id=app_installation.id, app_id=app.id
+        )
+
+
+@patch("saleor.app.installation_utils.fetch_icon_image")
+def test_fetch_brand_data_task_saving_brand_data(
+    mock_fetch_icon_image, app_installation, app, media_root
+):
+    # given
+    brand_data = {"logo": {"default": "https://example.com/logo.png"}}
+    fake_img_content = b"these are bytes"
+    logo_img = ContentFile(fake_img_content, "logo.png")
+
+    def fake_fetch_icon_image(*args, **kwargs):
+        # AppInstallation deleted during brand data fetching
+        app_installation.delete()
+        return logo_img
+
+    mock_fetch_icon_image.side_effect = fake_fetch_icon_image
+
+    # when
+    fetch_brand_data_task(
+        brand_data, app_installation_id=app_installation.id, app_id=app.id
+    )
+
+    # then
+    app.refresh_from_db()
+    assert app.brand_logo_default.read() == fake_img_content
+
+
+@patch("saleor.app.installation_utils.AppInstallation.save", side_effect=DatabaseError)
+@patch("saleor.app.installation_utils.default_storage.delete")
+@patch("saleor.app.installation_utils.fetch_icon_image")
+def test_fetch_brand_data_task_saving_deleted_object(
+    mock_fetch_icon_image,
+    mock_storage_delete,
+    _mock_app_installation_save,
+    app_installation,
+    media_root,
+):
+    # given
+    brand_data = {"logo": {"default": "https://example.com/logo.png"}}
+    file_name = "logo.png"
+    mock_fetch_icon_image.return_value = ContentFile(b"these are bytes", file_name)
+
+    # when
+    fetch_brand_data_task(brand_data, app_installation_id=app_installation.id)
+
+    # then
+    mock_storage_delete.assert_called_once_with(
+        f"app-installation-brand-data/{file_name}"
+    )