+++ saleor/graphql/discount/mutations/sale/sale_update.py
@@ -4,19 +4,19 @@ from datetime import datetime
 import graphene
 import pytz
 
-from ....core.tracing import traced_atomic_transaction
-from ....discount import models
-from ....discount.utils import CATALOGUE_FIELDS, fetch_catalogue_info
-from ....permission.enums import DiscountPermissions
-from ....product.tasks import update_products_discounted_prices_of_catalogues_task
-from ...channel import ChannelContext
-from ...core import ResolveInfo
-from ...core.mutations import ModelMutation
-from ...core.types import DiscountError
-from ...plugins.dataloaders import get_plugin_manager_promise
-from ..types import Sale
+from .....core.tracing import traced_atomic_transaction
+from .....discount import models
+from .....discount.utils import CATALOGUE_FIELDS, fetch_catalogue_info
+from .....permission.enums import DiscountPermissions
+from .....product.tasks import update_products_discounted_prices_of_catalogues_task
+from ....channel import ChannelContext
+from ....core import ResolveInfo
+from ....core.mutations import ModelMutation
+from ....core.types import DiscountError
+from ....plugins.dataloaders import get_plugin_manager_promise
+from ...types import Sale
+from ..utils import convert_catalogue_info_to_global_ids
 from .sale_create import SaleInput
-from .utils import convert_catalogue_info_to_global_ids
 
 
 class SaleUpdate(ModelMutation):