+++ saleor/graphql/discount/mutations/voucher/voucher_add_catalogues.py
@@ -1,14 +1,14 @@
 import graphene
 
-from ....permission.enums import DiscountPermissions
-from ...channel import ChannelContext
-from ...core import ResolveInfo
-from ...core.descriptions import ADDED_IN_31
-from ...core.doc_category import DOC_CATEGORY_DISCOUNTS
-from ...core.types import BaseInputObjectType, DiscountError, NonNullList
-from ...plugins.dataloaders import get_plugin_manager_promise
-from ..types import Voucher
-from .sale_base_discount_catalogue import BaseDiscountCatalogueMutation
+from .....permission.enums import DiscountPermissions
+from ....channel import ChannelContext
+from ....core import ResolveInfo
+from ....core.descriptions import ADDED_IN_31
+from ....core.doc_category import DOC_CATEGORY_DISCOUNTS
+from ....core.types import BaseInputObjectType, DiscountError, NonNullList
+from ....plugins.dataloaders import get_plugin_manager_promise
+from ...types import Voucher
+from ..sale.sale_base_discount_catalogue import BaseDiscountCatalogueMutation
 
 
 class CatalogueInput(BaseInputObjectType):