+++ saleor/graphql/discount/mutations/voucher/voucher_channel_listing_update.py
@@ -4,19 +4,19 @@ from typing import Dict, List
 import graphene
 from django.core.exceptions import ValidationError
 
-from ....core.tracing import traced_atomic_transaction
-from ....discount import DiscountValueType, models
-from ....discount.error_codes import DiscountErrorCode
-from ....permission.enums import DiscountPermissions
-from ...channel import ChannelContext
-from ...channel.mutations import BaseChannelListingMutation
-from ...core import ResolveInfo
-from ...core.doc_category import DOC_CATEGORY_DISCOUNTS
-from ...core.scalars import PositiveDecimal
-from ...core.types import BaseInputObjectType, DiscountError, NonNullList
-from ...core.validators import validate_price_precision
-from ...plugins.dataloaders import get_plugin_manager_promise
-from ..types import Voucher
+from .....core.tracing import traced_atomic_transaction
+from .....discount import DiscountValueType, models
+from .....discount.error_codes import DiscountErrorCode
+from .....permission.enums import DiscountPermissions
+from ....channel import ChannelContext
+from ....channel.mutations import BaseChannelListingMutation
+from ....core import ResolveInfo
+from ....core.doc_category import DOC_CATEGORY_DISCOUNTS
+from ....core.scalars import PositiveDecimal
+from ....core.types import BaseInputObjectType, DiscountError, NonNullList
+from ....core.validators import validate_price_precision
+from ....plugins.dataloaders import get_plugin_manager_promise
+from ...types import Voucher
 
 
 class VoucherChannelListingAddInput(BaseInputObjectType):