+++ saleor/graphql/discount/mutations/voucher/voucher_remove_catalogues.py
@@ -1,9 +1,9 @@
-from ....permission.enums import DiscountPermissions
-from ...core import ResolveInfo
-from ...core.doc_category import DOC_CATEGORY_DISCOUNTS
-from ...core.types import DiscountError
-from ...plugins.dataloaders import get_plugin_manager_promise
-from ..types import Voucher
+from .....permission.enums import DiscountPermissions
+from ....core import ResolveInfo
+from ....core.doc_category import DOC_CATEGORY_DISCOUNTS
+from ....core.types import DiscountError
+from ....plugins.dataloaders import get_plugin_manager_promise
+from ...types import Voucher
 from .voucher_add_catalogues import VoucherBaseCatalogueMutation